+++ NEWS.md
@@ -1,3 +1,5 @@
+# fable (development version)
+
 # fable 0.4.1
 
 ## Bug fixes