+++ NEWS.md
@@ -10,6 +10,7 @@
 * Fixed `VARIMA()` not working with integer data.
 * More robust calculations of `ARIMA()` AR and MA roots (#417).
 * Fixed `refit()` method for `ETS()` models applying transformations twice (#407).
+* Fixed recycling warning in `generate()` method for `SNAIVE()` models (#436).
 
 # fable 0.4.1