+++ NEWS.md
@@ -1,4 +1,4 @@
-# fable (development version)
+# fable 0.5.0
 
 ## New features